+++ particle_random_walk/INCLUDE/cluster.h
@@ -22,7 +22,7 @@ public:
 	~Cluster();
 
 	inline float get_potential(int x, int y) { return cluster_field[y][x]; }
-	inline float get_element(int x, int y) { return cluster_elements[y][x]; }
+	inline int get_element(int x, int y) { return cluster_elements[y][x]; }
 	inline int get_width() { return width; }
 	inline int get_height() { return height; }
 	inline float get_sigma() { return sigma; }
